Crate pandoc_ast

Structs

Citation

Format
Formats for raw blocks

Map
A map based on a B-Tree.

Pandoc
the root object of a pandoc document

Enums

Alignment
Alignment of a table column.

Block
Structured text like tables and lists

CitationMode

Inline
a single formatting item like bold, italic or hyperlink

ListNumberDelim
Delimiter of list numbers.

ListNumberStyle
Style of list numbers.

MathType
Type of math element (display or inline).

MetaValue

QuoteType
Type of quotation marks to use in Quoted inline.

Functions

filter
deserialized a json string to a Pandoc object, passes it to the closure/function and serializes the result back into a string

Type Definitions

Attr
Attributes: identifier, classes, key-value pairs

Double

Int

ListAttributes

TableCell
Table cells are list of Blocks

Target
Link target (URL, title).