Enum pandoc_ast::Block

pub enum Block {
    Plain {
        c: Vec<Inline>,
    },
    Para {
        c: Vec<Inline>,
    },
    LineBlock {
        c: Vec<Vec<Inline>>,
    },
    CodeBlock {
        c: (Attr, String),
    },
    RawBlock {
        c: (Format, String),
    },
    BlockQuote {
        c: Vec<Block>,
    },
    OrderedList {
        c: (ListAttributes, Vec<Vec<Block>>),
    },
    BulletList {
        c: Vec<Vec<Block>>,
    },
    DefinitionList {
        c: Vec<(Vec<Inline>, Vec<Vec<Block>>)>,
    },
    Header {
        c: (Int, Attr, Vec<Inline>),
    },
    HorizontalRule,
    Table {
        c: (Vec<Inline>, Vec<Alignment>, Vec<Double>, Vec<TableCell>, Vec<Vec<TableCell>>),
    },
    Div {
        c: (Attr, Vec<Block>),
    },
    Null,
}

Structured text like tables and lists

Variants
Plain

Plain text, not a paragraph

Fields of Plain

c: Vec<Inline>

Para

Paragraph

Fields of Para

c: Vec<Inline>

LineBlock

Multiple non-breaking lines

Fields of LineBlock

c: Vec<Vec<Inline>>

CodeBlock

Code block (literal) with attributes

Fields of CodeBlock

c: (Attr, String)

RawBlock
Fields of RawBlock

c: (Format, String)

BlockQuote

Block quote (list of blocks)

Fields of BlockQuote

c: Vec<Block>

OrderedList

Ordered list (attributes and a list of items, each a list of blocks)

Fields of OrderedList

c: (ListAttributes, Vec<Vec<Block>>)

BulletList

Bullet list (list of items, each a list of blocks)

Fields of BulletList

c: Vec<Vec<Block>>

DefinitionList

Definition list Each list item is a pair consisting of a term (a list of inlines) and one or more definitions (each a list of blocks)

Fields of DefinitionList

c: Vec<(Vec<Inline>, Vec<Vec<Block>>)>

Header

Header - level (integer) and text (inlines)

Fields of Header

c: (Int, Attr, Vec<Inline>)

HorizontalRuleTable

Table, with caption, column alignments (required), relative column widths (0 = default), column headers (each a list of blocks), and rows (each a list of lists of blocks)

Fields of Table

c: (Vec<Inline>, Vec<Alignment>, Vec<Double>, Vec<TableCell>, Vec<Vec<TableCell>>)

Div

Generic block container with attributes

Fields of Div

c: (Attr, Vec<Block>)

Null

Nothing